oneself-movies/requirements.txt
streamlit
pandas
numpy
scikit-learn
mysql-connector-python
requests
